+++ src/main/java/com/example/techit7/user/dto/UserCreateResponseDto.java
@@ -1,6 +1,6 @@
 package com.example.techit7.user.dto;
-import com.example.techit7.global.dto.GlobalResponseDto;
+import com.example.techit7.global.dto.GlobalResponse;
 
-public class UserCreateResponseDto extends GlobalResponseDto {
+public class UserCreateResponseDto extends GlobalResponse {
 
 }
