+++ src/main/java/com/example/techit7/user/service/UserServiceImpl.java
@@ -1,14 +1,12 @@
 package com.example.techit7.user.service;
 
 
-import com.example.techit7.global.dto.GlobalResponseDto;
+import com.example.techit7.global.dto.GlobalResponse;
 import com.example.techit7.user.dto.UserCreateRequestDto;
 import com.example.techit7.user.dto.UserCreateResponseDto;
 import com.example.techit7.user.entity.SiteUser;
 import com.example.techit7.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.crypto.bcrypt.BCrypt;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -32,7 +30,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public GlobalResponseDto postUser(UserCreateRequestDto userCreateRequestDto){
+    public GlobalResponse postUser(UserCreateRequestDto userCreateRequestDto){
         SiteUser siteUser = SiteUser.builder()
                 .username(userCreateRequestDto.getUsername())
                 .password(encoder.encode(userCreateRequestDto.getPassword1()))
@@ -56,7 +54,7 @@ public class UserServiceImpl implements UserService {
 
     }
     @Override
-    public GlobalResponseDto updateUserPassword(String username, UserCreateRequestDto dto) {
+    public GlobalResponse updateUserPassword(String username, UserCreateRequestDto dto) {
         SiteUser siteUser = findByUsername(username);
 
         userRepository.save(
@@ -64,12 +62,12 @@ public class UserServiceImpl implements UserService {
                         .password(encoder.encode(dto.getPassword1()))
                         .build());
 
-        return GlobalResponseDto.of("200","유저 비밀번호 변경 성공");
+        return GlobalResponse.of("200","유저 비밀번호 변경 성공");
     }
 
 
     @Override
-    public GlobalResponseDto updateUser(String username, UserCreateRequestDto dto) {
+    public GlobalResponse updateUser(String username, UserCreateRequestDto dto) {
         SiteUser siteUser = findByUsername(username);
 
         userRepository.save(
@@ -78,7 +76,7 @@ public class UserServiceImpl implements UserService {
                         .email(dto.getEmail())
                         .build());
 
-        return GlobalResponseDto.of("200","유저정보 변경 성공");
+        return GlobalResponse.of("200","유저정보 변경 성공");
     }
 
 }
